requirements.txt
django>=1.5
south
django-grappelli>=2.4.5
django-filebrowser>=3.5
Pillow

# django-epiceditor
markdown
django-pagedown
django-mptt
